Move common code to an httputil package
